Refactor BREP shape rendering into shared helper
