+++ lib/drawer/shapes/index.ts
@@ -1,21 +1,26 @@
-export { drawCircle, type DrawCircleParams } from "./circle"
-export { drawRect, type DrawRectParams } from "./rect"
-export { drawOval, type DrawOvalParams } from "./oval"
-export { drawPill, type DrawPillParams } from "./pill"
-export { drawPolygon, type DrawPolygonParams } from "./polygon"
-export { drawLine, type DrawLineParams } from "./line"
-export { drawPath, type DrawPathParams } from "./path"
-export { drawArrow, type DrawArrowParams } from "./arrow"
+export { type DrawArrowParams, drawArrow } from "./arrow"
 export {
-  drawDimensionLine,
+  addBrepRingToPath,
+  addBrepShapeToPath,
+  type BrepShape,
+} from "./brep"
+export { type DrawCircleParams, drawCircle } from "./circle"
+export {
   type DrawDimensionLineParams,
+  drawDimensionLine,
 } from "./dimension-line"
+export { type DrawLineParams, drawLine } from "./line"
+export { type DrawOvalParams, drawOval } from "./oval"
+export { type DrawPathParams, drawPath } from "./path"
+export { type DrawPillParams, drawPill } from "./pill"
+export { type DrawPolygonParams, drawPolygon } from "./polygon"
+export { type DrawRectParams, drawRect } from "./rect"
 export {
-  drawText,
-  type DrawTextParams,
-  getAlphabetLayout,
-  strokeAlphabetText,
-  getTextStartPosition,
   type AlphabetLayout,
   type AnchorAlignment,
+  type DrawTextParams,
+  drawText,
+  getAlphabetLayout,
+  getTextStartPosition,
+  strokeAlphabetText,
 } from "./text"
